+++ CHANGES.md
@@ -3,4 +3,11 @@
 16:51 Adding $ and $[0-9]* bash like arguments  
 17:05 Adding [ and ]  
 18:57 Adding single quoted strings, ampersand, hash and pipe  
-18:59 Adding option to tokenize spaces  
+18:59 Adding option to tokenize spaces
+
++ 04/04/2021
+04:59 Adding < > <= and >=  
+
+
+
+